only save file when it is modified
git-svn-id: https://repos.nextthought.com/svn/nti-svn/AoPS/trunk/src/main/python@286 3dfb626b-079f-4ab1-a9f6-4654e5377508

+++ tociconsetter.py
@@ -13,14 +13,16 @@ def main(args):
 	toc = dom.getElementsByTagName("toc");
 	if (toc):
 		current = 0
+		modified = False
 		for node in toc[0].childNodes:
 			if (node.nodeType == Node.ELEMENT_NODE and node.localName == 'topic'):
 				current += 1
-				handleTopic(node, current)
+				modified = handleTopic(node, current) | modified
 				
-		tempfile = toXml(dom)
-		os.remove(file)
-		os.rename(tempfile, file)
+		if (modified):
+			tempfile = toXml(dom)
+			os.remove(file)
+			os.rename(tempfile, file)
 
 def handleTopic(topic, current):
 	attributes = topic.attributes
@@ -28,6 +30,8 @@ def handleTopic(topic, current):
 		if (not hasIcon(attributes)):
 			s = 'C' + str(current) + '.png' 
 			attributes['icon'] = 'icons/chapters/' + s
+			return True
+	return False
 		
 def hasIcon(attributes):
 	return (attributes and attributes.get('icon'))
